Finnaly Medicine test is here.... Earlier it was failing due to my lack of understanding of tests.... I came to know that setUp method executes everytime for every test function... ANd I also updated the serializer so that only staff's can create medicines....
